requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
propcache==0.3.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.40
starlette==0.46.2
Telethon==1.39.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
WTForms==3.1.2
yarl==1.20.0
